requirements.txt
uWSGI
django==3.1.2
emoji
regex
psycopg2-binary
Pillow
djangorestframework
django-cors-headers
requests

